+++ test/feature/Engine-feature.js
@@ -1,4 +1,4 @@
-import { EventEmitter } from 'events';
+import { EventEmitter } from 'node:events';
 import { Engine } from '../../src/index.js';
 import * as factory from '../helpers/factory.js';
 import { camundaBpmnModdle as camunda } from '../helpers/testHelpers.js';